@@ -39,10 +39,10 @@ set(CMAKE_INCLUDE_CURRENT_DIR ON)
 set(CMAKE_MACOSX_RPATH OFF)
 
 ## #############################################################################
-## enable c++ 11
+## enable c++ 17
 ## #############################################################################
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 ## #############################################################################
